code clean

+++ apps/rahat-community/src/app/groups/group.service.ts
@@ -169,51 +169,10 @@ export class GroupService {
     deleteBeneficiaryFlag: boolean,
     beneficiaryUuid: string[],
   ) {
-    // get relevant informationn from the group table
-    // const group = await this.findUnique(uuid);
-    // console.log('Group=>', group);
-    // if (!group) throw new Error('Group not found');
-
-    // await this.prisma.$transaction(async (prisma) => {
-    //   for (const benefUid of beneficiaryUuid) {
-    //     const beneficiarygroup = await prisma.beneficiaryGroup.findFirst({
-    //       where: {
-    //         beneficiaryUID: benefUid,
-    //         groupUID: uuid,
-    //       },
-    //     });
-
-    //     if (beneficiarygroup) {
-    //       await prisma.beneficiaryGroup.updateMany({
-    //         where: {
-    //           beneficiaryUID: benefUid,
-    //           groupUID: uuid,
-    //         },
-    //         data: {
-    //           groupUID: null,
-    //         },
-    //       });
-
-    //       if (deleteBeneficiaryFlag) {
-    //         await prisma.beneficiary.update({
-    //           where: {
-    //             uuid: benefUid,
-    //           },
-    //           data: {
-    //             archived: true,
-    //           },
-    //         });
-    //       }
-    //     }
-    //   }
-    // });
-
-    console.log({ deleteBeneficiaryFlag });
-
+    const group = await this.findUnique(uuid);
+    if (!group) throw new Error('Group not found');
     if (group) {
       for (const item of beneficiaryUuid) {
-        // Delete from the group table (tbl_beneficiary_groups)
-        console.log({ item });
         await this.beneficaryGroupService.removeBeneficiaryFromGroup(
           item,
           uuid,
@@ -226,6 +185,7 @@ export class GroupService {
         }
       }
     }
+
     return 'Beneficiary removed successfully!';
   }
